fix: repair broken layout and refactor theme files
