+++ applications/tari_dan_wallet_web_ui/src/Components/SecondaryHeading.tsx
@@ -20,7 +20,7 @@
 //  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 //  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-import theme from '../theme';
+import theme from '../theme/theme';
 
 interface Props {
   children: string;
