+++ applications/tari_dan_wallet_web_ui/src/theme/theme.css
@@ -1,0 +1,65 @@
+@font-face {
+  src: url('./fonts/AvenirLTStd-Book.otf');
+  font-family: 'AvenirRegular';
+}
+
+@font-face {
+  src: url('./fonts/AvenirLTStd-Medium.otf');
+  font-family: 'AvenirMedium';
+}
+
+@font-face {
+  src: url('./fonts/AvenirLTStd-Heavy.otf');
+  font-family: 'AvenirHeavy';
+}
+
+body {
+  margin: 0;
+  font-family: 'AvenirMedium', sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  background-color: #f5f5f7;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
+}
+
+.add-confirm-form {
+  display: flex;
+  flex-direction: column;
+  align-content: center;
+  gap: 10px;
+}
+
+.flex-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: stretch;
+  width: 100%;
+  gap: 10px;
+}
+
+.heading-menu {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  gap: 5px;
+}
+
+@media screen and (max-width: 767px) {
+  .flex-container > * {
+    width: 100%;
+  }
+}
+
+@media screen and (min-width: 768px) {
+  .add-confirm-form {
+    flex-direction: row;
+  }
+  .flex-container {
+    flex-direction: row;
+  }
+}
